WDR: Basisklasse für Radiosender erstellt
