+++ src/main/java/de/mediathekview/mserver/crawler/wdr/Wdr4Crawler.java
@@ -1,33 +1,19 @@
 package de.mediathekview.mserver.crawler.wdr;
 
-import de.mediathekview.mlib.daten.Film;
 import de.mediathekview.mlib.daten.Sender;
 import de.mediathekview.mlib.messages.listener.MessageListener;
 import de.mediathekview.mserver.base.config.MServerConfigManager;
-import de.mediathekview.mserver.base.messages.ServerMessages;
-import de.mediathekview.mserver.crawler.basic.AbstractCrawler;
 import de.mediathekview.mserver.crawler.basic.CrawlerUrlDTO;
-import de.mediathekview.mserver.crawler.basic.TopicUrlDTO;
-import de.mediathekview.mserver.crawler.wdr.tasks.WdrFilmDetailTask;
 import de.mediathekview.mserver.crawler.wdr.tasks.WdrRadioPageTask;
-import de.mediathekview.mserver.crawler.wdr.tasks.WdrTopicOverviewTask;
 import de.mediathekview.mserver.progress.listeners.SenderProgressListener;
-import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
-import java.time.temporal.ChronoUnit;
 import java.util.Collection;
 import java.util.Set;
 import java.util.concurrent.ConcurrentLinkedQueue;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.ForkJoinPool;
-import java.util.concurrent.RecursiveTask;
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
 
-public class Wdr4Crawler extends AbstractCrawler {
+public class Wdr4Crawler extends WdrRadioCrawlerBase {
   
-  private static final Logger LOG = LogManager.getLogger(Wdr4Crawler.class);  
-
   public Wdr4Crawler(ForkJoinPool aForkJoinPool, Collection<MessageListener> aMessageListeners, Collection<SenderProgressListener> aProgressListeners, MServerConfigManager rootConfig) {
     super(aForkJoinPool, aMessageListeners, aProgressListeners, rootConfig);
   }
@@ -38,47 +24,11 @@ public class Wdr4Crawler extends AbstractCrawler {
   }
 
   @Override
-  protected RecursiveTask<Set<Film>> createCrawlerTask() {
-    try {
-      ConcurrentLinkedQueue<TopicUrlDTO> shows = new ConcurrentLinkedQueue<>();
-      shows.addAll(getEntries());
-      
-      printMessage(ServerMessages.DEBUG_ALL_SENDUNG_FOLGEN_COUNT, getSender().getName(), shows.size());
-      getAndSetMaxCount(shows.size());
-
-      return new WdrFilmDetailTask(this, shows);
-    } catch (InterruptedException | ExecutionException ex) {
-      LOG.fatal("Exception in WDR crawler.", ex);
-    }
-    return null;
-  }
-  
-  private Set<TopicUrlDTO> getEntries() throws InterruptedException, ExecutionException {
+  protected Set<WdrTopicUrlDTO> getTopicOverviewPages() throws InterruptedException, ExecutionException {
     ConcurrentLinkedQueue<CrawlerUrlDTO> urlToCrawl = new ConcurrentLinkedQueue<>();
-    ConcurrentLinkedQueue<WdrTopicUrlDTO> letterPageEntries = new ConcurrentLinkedQueue<>();
-
     urlToCrawl.add(new CrawlerUrlDTO(WdrConstants.URL_RADIO_WDR4));
+    
     WdrRadioPageTask radioPageTask = new WdrRadioPageTask(this, urlToCrawl);
-    letterPageEntries.addAll(forkJoinPool.submit(radioPageTask).get());
-
-    WdrTopicOverviewTask overviewTask = new WdrTopicOverviewTask(this, letterPageEntries, 0);
-    Set<TopicUrlDTO> shows = forkJoinPool.submit(overviewTask).get();
-
-    printMessage(ServerMessages.DEBUG_ALL_SENDUNG_FOLGEN_COUNT, getSender().getName(), shows.size());
-    return shows;
+    return forkJoinPool.submit(radioPageTask).get();
   }
-  
-  private ConcurrentLinkedQueue<CrawlerUrlDTO> getDayUrls() {
-    final ConcurrentLinkedQueue<CrawlerUrlDTO> urls = new ConcurrentLinkedQueue<>();
-    for (int i = 0; 
-      i < crawlerConfig.getMaximumDaysForSendungVerpasstSection() + crawlerConfig.getMaximumDaysForSendungVerpasstSectionFuture(); 
-      i++) {
-      urls.add(new CrawlerUrlDTO(String.format(WdrConstants.URL_DAY, 
-        LocalDateTime.now()
-        .plus(crawlerConfig.getMaximumDaysForSendungVerpasstSectionFuture(), ChronoUnit.DAYS)
-        .minus(i, ChronoUnit.DAYS).format(DateTimeFormatter.ofPattern("ddMMyyyy")))));
-    }
-
-    return urls;
-  }  
 }
